Replace warp fee controller with splitter
