--- .changeset/warp-fee-controller.md
@@ -1,6 +1,0 @@
----
-'@hyperlane-xyz/core': minor
-'@hyperlane-xyz/sdk': minor
----
-
-Added the warp fee controller contract and SDK helper for collecting warp route fees through ICAs and donating them to hub-chain LP routers.
